+++ blocks/cross-site-search/index.asset.php
@@ -1,1 +1,0 @@
-<?php return array( 'dependencies' => array( 'wp-blocks', 'wp-element', 'wp-block-editor', 'wp-components', 'wp-i18n' ), 'version' => '0.2.0' );