+++ blocks/cross-site-search/view.asset.php
@@ -1,1 +1,0 @@
-<?php return array( 'dependencies' => array(), 'version' => '0.2.0' );